A lot of code refactoring and simplification, commands are now generic and can be created from nowhere. File downloader (probably) fully working, including multifile download and cancellation.

+++ BlackHole.Slave/FileHelper.cs
@@ -21,7 +21,7 @@ namespace BlackHole.Slave
         /// <param name="currentPart"></param>
         /// <param name="path"></param>
         /// <returns></returns>
-        public static DownloadedFilePartMessage DownloadFilePart(int id, long currentPart, string path)
+        public static DownloadedFilePartMessage DownloadFilePart(long id, long currentPart, string path)
             => CommonHelper.DownloadFilePart(id, currentPart, path);
 
         /// <summary>
